requirements.txt
lance==1.2.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-experimental==0.3.4
langchain-text-splitters==0.3.8
openai==1.90.0
langchain_openai==0.3.24
pynvml==12.0.0
ollama==0.5.1
langgraph==0.4.8
easygui==0.98.3
pypandoc_binary==1.15
langchain_nomic==0.1.4
ruamel_yaml==0.18.14
pydantic==2.12
pydantic_core==2.41.1
typing_extensions==4.15.0
typing_inspection==0.4.2
sdg-hub[examples]==0.3.1
fastmcp>=2.0.0
uvicorn
starlette
graphrag
